mvp/backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
opencv-python==4.8.0.74
mediapipe==0.10.1
openai==0.28.0
python-dotenv==1.0.0
numpy==1.24.3
Werkzeug==2.3.6
speech-recognition==3.10.0
pyaudio==0.2.11
moviepy==1.0.3
